Remove online sign and move sign tests into wallet
